requirements.txt
click>=8.0.0
jinja2>=3.0.0
pyyaml>=5.0.0
pytest>=7.0.0
pytest-cov>=3.0.0
